examples/trig/requirements.txt

# в build_apk.sh при сборке "p4a apk" есть аргумент requirements, туда тоже добавить верхний блок

pygame==2.6.1        # making multimedia apps like games built on top of the SDL library. C, Python, Native, OpenGL
numpy==2.2.5         # fundamental package for scientific computing with Python
ecs-pattern==1.3.0   # Implementation of the ECS pattern for creating games.
pyjnius==1.6.1       # Access Java classes from Python

pip==25.1.1
setuptools==80.1.0
wheel==0.46.1
ruff==0.11.11
